create upgrade conftest
